Update handbook files for placeholder rename
