+++ lib/handbook/addon/docs/components/osf-placeholder/-components/circle/template.hbs
@@ -1,9 +1,9 @@
 <DocsDemo as |demo|>
-    <demo.example @name='placeholder.circle.hbs'>
+    <demo.example @name='osf-placeholder.circle.hbs'>
         <OsfPlaceholder as |placeholder|>
             <placeholder.circle />
         </OsfPlaceholder>
     </demo.example>
 
-    <demo.snippet @name='placeholder.circle.hbs' />
+    <demo.snippet @name='osf-placeholder.circle.hbs' />
 </DocsDemo>
