+++ lib/handbook/addon/docs/components/osf-placeholder/template.md
@@ -3,4 +3,4 @@
 Placeholders to be used for when loading lists.
 
 ## Circle
-{{docs/components/placeholder/-components/circle}}
+{{docs/components/osf-placeholder/-components/circle}}
