+++ lib/handbook/addon/docs/template.hbs
@@ -36,7 +36,7 @@
         <nav.item @label='osf-link' @route='docs.components.osf-link' />
         <nav.item @label='osf-dialog' @route='docs.components.osf-dialog' />
         <nav.item @label='panel' @route='docs.components.panel' />
-        <nav.item @label='placeholder' @route='docs.components.placeholder' />
+        <nav.item @label='osf-placeholder' @route='docs.components.osf-placeholder' />
         <nav.item @label='schema-block-group-renderer' @route='docs.components.schema-block-group-renderer' />
         <nav.item @label='tags-widget' @route='docs.components.tags-widget' />
         <nav.item @label='validated-model-form' @route='docs.components.validated-model-form' />
