+++ lib/handbook/addon/routes.js
@@ -44,7 +44,7 @@ export default buildRoutes(function() {
             this.route('osf-layout');
             this.route('osf-link');
             this.route('panel');
-            this.route('placeholder');
+            this.route('osf-placeholder');
             this.route('schema-chunk');
             this.route('subject-widget');
             this.route('schema-block-group-renderer');
