+++ lib/osf-components/addon/components/osf-placeholder/component.ts
@@ -4,5 +4,5 @@ import { layout } from 'ember-osf-web/decorators/component';
 import template from './template';
 
 @layout(template)
-export default class Placeholder extends Component {
+export default class OsfPlaceholder extends Component {
 }
