+++ lib/osf-components/addon/components/osf-placeholder/template.hbs
@@ -1,3 +1,3 @@
 {{yield (hash
-    circle=(component 'placeholder/circle')
+    circle=(component 'osf-placeholder/circle')
 )}}
