+++ lib/osf-components/app/components/osf-placeholder/circle/component.js
@@ -1,1 +1,1 @@
-export { default } from 'osf-components/components/placeholder/circle/component';
+export { default } from 'osf-components/components/osf-placeholder/circle/component';
