+++ lib/osf-components/app/components/osf-placeholder/component.js
@@ -1,1 +1,1 @@
-export { default } from 'osf-components/components/placeholder/component';
+export { default } from 'osf-components/components/osf-placeholder/component';
